Moved UserController to the right place.
