+++ pms-webserver/src/main/java/de/hfu/pms/service/UserService.java
@@ -1,9 +1,7 @@
 package de.hfu.pms.service;
 
-import de.hfu.pms.dao.UserDao;
 import de.hfu.pms.model.UserRole;
 import de.hfu.pms.shared.dto.UserDto;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 @Service
